+++ doc/Changelog.md
@@ -5,6 +5,8 @@ Changelog {#Changelog}
 
 ### 0.11.0 (git master)
 
+* [103](https://github.com/BlueBrain/Deflect/pull/103):
+  DesktopStreamer: prevent AppNap of being re-enabled automatically
 * [102](https://github.com/BlueBrain/Deflect/pull/102):
   DeflectQt: Continue rendering & streaming after updates for a while to
   compensate for running animations, fix spurious missing event handling